[openresty] Hardcode embedded libcrytpo and libssl shared library for build
The Openresty build system for some reason is not picking up libcrytpo
and libssl from the embedded libraries.

A patch is applied to the Nginx autoconf to link directly to the
embedded libcrytpo and libssl instead of relying on the linker to find
it.

However, this is not the ideal solution. The patch relies on having
install_dir being fixed in place. Further, it does not solve the
underlying problem in Nginx's build system.